[OPIK-7791] [QA] hoist the python metric sources into core/metrics
buildControlMetric (python-metric-errors) and buildAlwaysScoresMetric
(sampling-rate) were byte-identical apart from the python class name, and
buildOptionalArgsMetric differed only in score()'s parameter list. One
buildConstantScoreMetric with a scoreArgs parameter covers all three.

Plain string builders rather than fixtures: they hold no state and need no
setup or teardown, so there is nothing for a fixture's use() to bracket. The
two evaluator-failure snippets move alongside them, keeping every constraint on
generated metric source (no extra BaseMetric imports, ScoreResult name is what
lands on the trace) documented in one place.

Addresses review feedback from @natagh23.
