+++ tests_end_to_end/e2e/core/metrics/index.ts
@@ -1,0 +1,5 @@
+export {
+  buildConstantScoreMetric,
+  buildSilentMetric,
+  buildUnparseableMetric,
+} from './python-metric-source';
